Move common wallet connect types to common folder and rename EVM-specific types
