+++ src/common/types.ts
@@ -1,3 +1,4 @@
+import WalletConnect from '@walletconnect/client'
 import {
   asArray,
   asEither,
@@ -82,3 +83,41 @@ export function asIntegerString(raw: unknown): string {
   }
   return clean
 }
+
+export const asWcProps = asObject({
+  uri: asString,
+  language: asMaybe(asString),
+  token: asMaybe(asString)
+})
+
+export type WcProps = ReturnType<typeof asWcProps>
+
+const asWcDappDetails = asObject({
+  peerId: asString,
+  peerMeta: asObject({
+    description: asString,
+    url: asString,
+    icons: asArray(asString),
+    name: asString
+  }),
+  chainId: asOptional(asNumber, 1)
+})
+
+export type WcDappDetails = {
+  timeConnected: number
+} & ReturnType<typeof asWcDappDetails>
+
+export type Dapp = { timeConnected: number } & WcProps & WcDappDetails
+
+export interface WalletConnectors {
+  [uri: string]: {
+    connector: WalletConnect
+    wcProps: WcProps
+    dApp: WcDappDetails
+    walletId?: string
+  }
+}
+
+export const asWcSessionRequestParams = asObject({
+  params: asArray(asWcDappDetails)
+})
